Updated NVIM_CODE and README.md
NVIM_CODE

1. ALE
	Installed the ALE plugin to get inline syntax compilation
errors.

2. init.lua
	Added smartindent.

README.md

1. nvim_code documentation
	Started writing the documentation for nvim_code, for the moment
only including the 2 changes mention higher in this commit as it's the
only thing I've done here.

2. Syntax standard
	Changed the "Configuration" section of configs. I used to
write the settings in the way it was written in .vimrc for readability.
Changed my mind about that, code blocks are designated as 'lua' and
settings are written exactly the way they are in my init.lua.
	Thus changed the one for nvim_base, and wrote the one for
nvim_code like that.