Updated SHARED_CONFIG and README.md
SHARED_CONFIG

1. init.lua
	Completely removed every line of configuration from init.lua and
put them in their respective file corresponding to their plugins, in
/lua/config/. Created the nvim.lua (for vim settings) and keybinds.lua
(for global keybinds).

2. nvim-tree
	Changed keybinds for opening a file splitting
horizontally/vertically or in a new tab. Instead of CTRL combination,
just the letter corresponding to their action in the <C-w> pannel, alone,
to do that. Since 's' was already taken by "Run System", swapped it to
<C-s>. Deleted the others original keybinds (<C-v> and <C-t>).

README.md

1. nvim_base documentation
	Updated custom keybind specification for nvim-tree. Also
indicated the default keybind to show the list of all keybind. Decided
against including custom internal nvim-tree keybinds in my keybinds table,
since it wouldn't make sense to include these one and not the default
ones. Probably gonna change the keybind to show all keybind to '?',
because it makes more sense AND it justifies its inclusion in 'Custom'
Keybindings table. Which makes it not a big deal to not include other
keybinds there.